bug(Numbers/): Division not working but found the cause
When dealing with a bigger dividend than the divisor and one with its last entry zero, the length of the list of the dividend, and therefore the actual number, may have an extra zero. Attempted unsuccessfully to remove it using pop. Also, there are some minor errors when executing with the Wall flag.